Improve the "-d" and "-d -d" facilities.
When we close a CLI and it had fd# 0 and/or fd#1, reopen these
as /dev/null so the will not be reused for the CLI pipe to the
child on next restart, otherwise stdout/stderr output from the
manager would get sent there and confuse the clients CLI reader.

Don't double free a pointer to the CLI buffer.

Accept non-zero results from cli_readres() errors are non-fatal.

Use stderr more consistently for manager debugging.



git-svn-id: http://varnish.projects.linpro.no/svn/trunk@749 d4fa192b-c00b-0410-8231-f00ffab90ce4