Overhaul kqueue acceptor in light of todays learnings.
Use the pipe trick to inject sessions into the system, as far as I
can tell it is cheaper because of the low rate it happens and the
high rate of mutex operations avoided.

Ignore the timer event, but purge the list every time we wake up
to reduce lumpyness of timeout'ing.

Centralize the polling of a session so we don't have the same two
messages spread out all over the place.

Centralize the acceptor thread and send things directly to the worker
thread, leaving only the session-herder in the split out files.

poll & epoll not yet updated accordingly.



git-svn-id: http://varnish.projects.linpro.no/svn/trunk@875 d4fa192b-c00b-0410-8231-f00ffab90ce4