Reduce traffic on the wrk_mtx, saving some syscalls along the way:
Make a thread-reaper-thread which examines the tail end of the queue
and kicks threads which are too old so the wake up and die.

This allows us to assume that a thread on the free queue is always
waiting on its condvar, so we don't need to hold the mutex when we
signal the condvar.

As a result, the woken thread stands a chance to grab the mutex
and a little song and dance between the two threads is avoided.



git-svn-id: http://varnish.projects.linpro.no/svn/trunk@929 d4fa192b-c00b-0410-8231-f00ffab90ce4