Add a convenience function more for backend methods:
VBE_CheckFd(): Check that a filedescriptor is reusable.

Right now we simply poll it with a zero timeout, and if there are
any events, we can't reuse it.

This check may need refinement down the road.

One option would be to attempt to write a CRNL onto the fd and see
that it works.



git-svn-id: http://varnish.projects.linpro.no/svn/trunk@1965 d4fa192b-c00b-0410-8231-f00ffab90ce4