Start pulling ESI:include stuff over from my private tree.
This is the meat of the XML-dissecting code.

In theory the task is pretty simple, but it gets complicated by the fact
that the XML input can be split over multiple pieces of storage, with
an element we care about straddling a boundary.

I have not decided what to do about ESI violations or XML syntax errors,
SHMlogging them, possibly controlled by a parameter might make sense.
For now mark them in the source.



git-svn-id: http://varnish.projects.linpro.no/svn/trunk@2108 d4fa192b-c00b-0410-8231-f00ffab90ce4