Improve readability, such as it is.
Allow the user to disable specific acceptor mechanisms (e.g. do not
use epoll even though it is available).

Don't look for kqueue on systems where we know it doesn't work.

Bail if no acceptor mechanism was found.

Bail if no curses or ncurses was found.

Warn the user if SO_{RCV,SND}TIMEO are non-functional.


git-svn-id: http://varnish.projects.linpro.no/svn/trunk@2382 d4fa192b-c00b-0410-8231-f00ffab90ce4